+++ engine/apps/alerts/tests/test_notify_ical_schedule_shift.py
@@ -100,7 +100,7 @@ def test_next_shift_notification_long_shifts(
 
     with patch("apps.alerts.tasks.notify_ical_schedule_shift.datetime", Mock(wraps=datetime)) as mock_datetime:
         mock_datetime.datetime.now.return_value = datetime.datetime(2021, 9, 29, 12, 0, tzinfo=pytz.UTC)
-        with patch("apps.slack.client.SlackClientWithErrorHandling.chat_postMessage") as mock_slack_api_call:
+        with patch("apps.slack.client.SlackClient.chat_postMessage") as mock_slack_api_call:
             notify_ical_schedule_shift(ical_schedule.pk)
 
     slack_blocks = mock_slack_api_call.call_args_list[0][1]["blocks"]
@@ -204,7 +204,7 @@ def test_overrides_changes_no_current_no_triggering_notification(
     schedule.prev_ical_file_overrides = ical_before
     schedule.save()
 
-    with patch("apps.slack.client.SlackClientWithErrorHandling.chat_postMessage") as mock_slack_api_call:
+    with patch("apps.slack.client.SlackClient.chat_postMessage") as mock_slack_api_call:
         notify_ical_schedule_shift(schedule.pk)
 
     assert not mock_slack_api_call.called
@@ -252,7 +252,7 @@ def test_no_changes_no_triggering_notification(
     schedule.empty_oncall = False
     schedule.save()
 
-    with patch("apps.slack.client.SlackClientWithErrorHandling.chat_postMessage") as mock_slack_api_call:
+    with patch("apps.slack.client.SlackClient.chat_postMessage") as mock_slack_api_call:
         notify_ical_schedule_shift(schedule.pk)
 
     assert not mock_slack_api_call.called
@@ -300,7 +300,7 @@ def test_current_shift_changes_trigger_notification(
     schedule.empty_oncall = False
     schedule.save()
 
-    with patch("apps.slack.client.SlackClientWithErrorHandling.chat_postMessage") as mock_slack_api_call:
+    with patch("apps.slack.client.SlackClient.chat_postMessage") as mock_slack_api_call:
         notify_ical_schedule_shift(schedule.pk)
 
     assert mock_slack_api_call.called
@@ -364,7 +364,7 @@ def test_current_shift_changes_swap_split(
     schedule.empty_oncall = False
     schedule.save()
 
-    with patch("apps.slack.client.SlackClientWithErrorHandling.chat_postMessage") as mock_slack_api_call:
+    with patch("apps.slack.client.SlackClient.chat_postMessage") as mock_slack_api_call:
         notify_ical_schedule_shift(schedule.pk)
 
     text_block = mock_slack_api_call.call_args_list[0][1]["blocks"][1]["text"]["text"]
@@ -433,7 +433,7 @@ def test_next_shift_changes_no_triggering_notification(
     on_call_shift_2.add_rolling_users([[user2]])
     schedule.refresh_ical_file()
 
-    with patch("apps.slack.client.SlackClientWithErrorHandling.chat_postMessage") as mock_slack_api_call:
+    with patch("apps.slack.client.SlackClient.chat_postMessage") as mock_slack_api_call:
         notify_ical_schedule_shift(schedule.pk)
 
     assert not mock_slack_api_call.called
@@ -500,7 +500,7 @@ def test_lower_priority_changes_no_triggering_notification(
     on_call_shift_2.add_rolling_users([[user2]])
     schedule.refresh_ical_file()
 
-    with patch("apps.slack.client.SlackClientWithErrorHandling.chat_postMessage") as mock_slack_api_call:
+    with patch("apps.slack.client.SlackClient.chat_postMessage") as mock_slack_api_call:
         notify_ical_schedule_shift(schedule.pk)
 
     assert not mock_slack_api_call.called
@@ -630,7 +630,7 @@ def test_vtimezone_changes_no_triggering_notification(
     schedule.cached_ical_file_primary = ical_after
     schedule.save()
 
-    with patch("apps.slack.client.SlackClientWithErrorHandling.chat_postMessage") as mock_slack_api_call:
+    with patch("apps.slack.client.SlackClient.chat_postMessage") as mock_slack_api_call:
         notify_ical_schedule_shift(schedule.pk)
 
     assert not mock_slack_api_call.called
@@ -687,7 +687,7 @@ def test_no_changes_no_triggering_notification_from_old_to_new_task_version(
     schedule.empty_oncall = False
     schedule.save()
 
-    with patch("apps.slack.client.SlackClientWithErrorHandling.chat_postMessage") as mock_slack_api_call:
+    with patch("apps.slack.client.SlackClient.chat_postMessage") as mock_slack_api_call:
         notify_ical_schedule_shift(schedule.pk)
 
     assert not mock_slack_api_call.called
@@ -749,7 +749,7 @@ def test_current_shift_changes_trigger_notification_from_old_to_new_task_version
     on_call_shift.add_rolling_users([[user2]])
     schedule.refresh_ical_file()
 
-    with patch("apps.slack.client.SlackClientWithErrorHandling.chat_postMessage") as mock_slack_api_call:
+    with patch("apps.slack.client.SlackClient.chat_postMessage") as mock_slack_api_call:
         notify_ical_schedule_shift(schedule.pk)
 
     assert mock_slack_api_call.called
@@ -814,7 +814,7 @@ def test_next_shift_notification_long_and_short_shifts(
     schedule.empty_oncall = False
     schedule.save()
 
-    with patch("apps.slack.client.SlackClientWithErrorHandling.chat_postMessage") as mock_slack_api_call:
+    with patch("apps.slack.client.SlackClient.chat_postMessage") as mock_slack_api_call:
         notify_ical_schedule_shift(schedule.pk)
 
     assert mock_slack_api_call.called