+++ engine/apps/schedules/ical_utils.py
@@ -29,6 +29,7 @@ from apps.schedules.constants import (
     ICAL_STATUS_CANCELLED,
     ICAL_SUMMARY,
     ICAL_UID,
+    RE_EVENT_UID_EXPORT,
     RE_EVENT_UID_V1,
     RE_EVENT_UID_V2,
     RE_PRIORITY,
@@ -365,18 +366,18 @@ def list_users_to_notify_from_ical_for_period(
 
 
 def get_oncall_users_for_multiple_schedules(
-    schedules: "OnCallScheduleQuerySet", events_datetime=None
+    schedules: typing.List["OnCallSchedule"], events_datetime=None
 ) -> typing.Dict["OnCallSchedule", UserQuerySet]:
     if events_datetime is None:
         events_datetime = datetime.datetime.now(timezone.utc)
 
     # Exit early if there are no schedules
-    if not schedules.exists():
+    if not schedules:
         return {}
 
     # Get on-call users
     oncall_users = {}
-    for schedule in schedules.all():
+    for schedule in schedules:
         # pass user list to list_users_to_notify_from_ical
         schedule_oncall_users = list_users_to_notify_from_ical(schedule, events_datetime=events_datetime)
         oncall_users.update({schedule.pk: schedule_oncall_users})
@@ -421,21 +422,25 @@ def parse_event_uid(string: str, sequence: str = None, recurrence_id: str = None
     if match:
         _, pk, _, _, source = match.groups()
     else:
-        # eventually this path would be automatically deprecated
-        # once all ical representations are refreshed
-        match = RE_EVENT_UID_V1.match(string)
+        match = RE_EVENT_UID_EXPORT.match(string)
         if match:
-            _, _, _, source = match.groups()
+            pk, _, _ = match.groups()
         else:
-            # fallback to use the UID string as the rotation ID
-            pk = string
-            # in ical imported calendars, sequence and/or recurrence_id
-            # distinguish main recurring event vs instance modification
-            # (see https://icalendar.org/iCalendar-RFC-5545/3-8-4-4-recurrence-id.html)
-            if sequence:
-                pk = f"{pk}_{sequence}"
-            if recurrence_id:
-                pk = f"{pk}_{recurrence_id}"
+            # eventually this path would be automatically deprecated
+            # once all ical representations are refreshed
+            match = RE_EVENT_UID_V1.match(string)
+            if match:
+                _, _, _, source = match.groups()
+            else:
+                # fallback to use the UID string as the rotation ID
+                pk = string
+                # in ical imported calendars, sequence and/or recurrence_id
+                # distinguish main recurring event vs instance modification
+                # (see https://icalendar.org/iCalendar-RFC-5545/3-8-4-4-recurrence-id.html)
+                if sequence:
+                    pk = f"{pk}_{sequence}"
+                if recurrence_id:
+                    pk = f"{pk}_{recurrence_id}"
 
     if source is not None:
         source = int(source)